--- a/Cargo.toml
+++ b/Cargo.toml
@@ -46,6 +46,17 @@ sysinfo = { version = "0.33.1", features = ["serde"] }
 indicatif = "0.17.8"
 console = "0.15.8"
 async-trait = "0.1.82"
+libc = "0.2.171"
+bincode = "1.3.3"
+object = "0.36.7"
+linux-perf-data = "0.11.0"
+debugid = "0.8.0"
+memmap2 = "0.9.5"
+nix = { version = "0.29.0", features = ["fs"] }
+futures = "0.3.31"
+
+[target.'cfg(target_os = "linux")'.dependencies]
+procfs = "0.17.0"
 
 [dev-dependencies]
 temp-env = { version = "0.3.6", features = ["async_closure"] }