--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "codspeed-runner"
-version = "3.5.0"
+version = "3.6.0-beta.2"
 edition = "2021"
 repository = "https://github.com/CodSpeedHQ/runner"
 publish = false