--- a/src/run/runner/helpers/mod.rs
+++ b/src/run/runner/helpers/mod.rs
@@ -2,3 +2,4 @@ pub mod env;
 pub mod get_bench_command;
 pub mod profile_folder;
 pub mod run_command_with_log_pipe;
+pub mod setup;