--- a/src/run/runner/valgrind/mod.rs
+++ b/src/run/runner/valgrind/mod.rs
@@ -1,4 +1,4 @@
 pub mod executor;
-mod helpers;
+pub mod helpers;
 mod measure;
 mod setup;