--- a/src/run/runner/wall_time/mod.rs
+++ b/src/run/runner/wall_time/mod.rs
@@ -1,1 +1,2 @@
 pub mod executor;
+pub mod perf;