TODOs: 
- [x] ~~Don't use max freq (or parse it and send it with the traces)~~ Actually doesn't matter. Higher freq, better callgraph. 
- [x] upload data + debug symbols to server
- [x] ~~use correct `codspeed` version (depends on https://github.com/CodSpeedHQ/codspeed-rust/pull/92)~~

Debug logs of successful run:
```
►►► Running the benchmarks
[DEBUG::codspeed::run::runner::helpers::setup] Running command: sudo sysctl -w kernel.kptr_restrict=0
[DEBUG::codspeed::run::runner::helpers::setup] Running command: sudo sysctl -w kernel.perf_event_paranoid=1
[DEBUG::codspeed::run::runner::wall_time::executor] cmd: ARCH="x86_64" CODSPEED_ENV="runner" CODSPEED_PROFILE_FOLDER="/tmp/profile.riowzUmLRF.out" CODSPEED_RUNNER_MODE="walltime" PYTHONHASHSEED="0" "sh" "-c" "perf record --data --freq=1000 --switch-output --control=fifo:/tmp/codspeed_perf.ctl.fifo,/tmp/codspeed_perf.ack.fifo --delay=-1 -g --output=/tmp/.tmpeI44Iq/perf.datao1Dpk6 -- cargo-codspeed run"
[DEBUG::codspeed::run::runner::wall_time::executor] Perf PID: 527912
Events disabled
[cargo-codspeed] Measurement mode: Walltime

Collected 1 benchmark suite(s) to run
Running graph-analysis my_bench
Timer precision: 13 ns
my_bench        fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ fibonacci                  │               │               │               │         │
Events enabled
[ perf record: dump data: Woken up 1 times ]
[ perf record: Dump /tmp/.tmpeI44Iq/perf.datao1Dpk6.2025032512120419 ]
Events disabled
│  ╰─ 32        4.563 ms      │ 7.109 ms      │ 4.881 ms      │ 5.07 ms       │ 100     │ 100
╰─ other_bench                │               │               │               │         │
Events enabled
[ perf record: dump data: Woken up 2 times ]
[ perf record: Dump /tmp/.tmpeI44Iq/perf.datao1Dpk6.2025032512120472 ]
Events disabled
   ╰─ 50        50.06 ms      │ 50.31 ms      │ 50.1 ms       │ 50.12 ms      │ 100     │ 100

Done running   ⠉ Running the benchmarks...                                                                                                                                   my_bench
Finished running 1 benchmark suite(s)
[ perf record: Woken up 2 times to write data ]
[ perf record: Dump /tmp/.tmpeI44Iq/perf.datao1Dpk6.2025032512120974 ]
[ perf record: Captured and wrote 0.034 MB /tmp/.tmpeI44Iq/perf.datao1Dpk6.<timestamp> ]
  ⠉ Running the benchmarks...                                                                                                                                                [src/run/runner/wall_time/executor.rs:117:9] &perf_files = [
    "/tmp/.tmpeI44Iq/perf.datao1Dpk6.2025032512120419",
    "/tmp/.tmpeI44Iq/perf.datao1Dpk6.2025032512120472",
    "/tmp/.tmpeI44Iq/perf.datao1Dpk6.2025032512120974",
]
```