@not-matthias 2025-05-07
@GuillaumeLagrange Only minor changes in this PR: Added conditional dependency of `procfs`, disabled `USE_PERF` by default, install perf on setup if not found.